pyproject.toml
[project]
name = "comma-auth"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.116.1",
    "google-auth>=2.40.3",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.2",
    "httpx>=0.28.1",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.20",
    "twilio>=9.6.5",
    "uvicorn>=0.35.0",
]
